+++ src/main/java/svenhjol/charmony/travel_journal/common/features/travel_journal/Handlers.java
@@ -1,6 +1,6 @@
 package svenhjol.charmony.travel_journal.common.features.travel_journal;
 
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.world.entity.player.Player;
 import svenhjol.charmony.core.base.Setup;
 import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.C2SPlayerSettings;
 import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.C2SSendBookmarkToPlayer;
@@ -19,49 +19,42 @@ public final class Handlers extends Setup<TravelJournal> {
         super(feature);
     }
 
-    public void handleSendBookmarkToPlayerPacket(C2SSendBookmarkToPlayer packet, ServerPlayNetworking.Context context) {
-        var player = context.player();
-        var server = player.server;
+    public void handleSendBookmarkToPlayerPacket(Player player, C2SSendBookmarkToPlayer payload) {
+        var server = player.getServer();
+        if (server == null) return;
 
-        server.execute(() -> {
-            var bookmark = packet.bookmark();
-            var photo = packet.photo();
-            var recipient = packet.recipient();
-            var sender = player.getScoreboardName();
+        var bookmark = payload.bookmark();
+        var photo = payload.photo();
+        var recipient = payload.recipient();
+        var sender = player.getScoreboardName();
 
-            // If target player has disabled the ability to receive bookmarks, return early.
-            if (!playerCanReceiveBookmarks.containsKey(recipient) || !playerCanReceiveBookmarks.get(recipient)) {
-                log().debug("Player " + recipient + " is not accepting bookmarks");
+        // If target player has disabled the ability to receive bookmarks, return early.
+        if (!playerCanReceiveBookmarks.containsKey(recipient) || !playerCanReceiveBookmarks.get(recipient)) {
+            log().debug("Player " + recipient + " is not accepting bookmarks");
+            return;
+        }
+
+        if (playerAllowBookmarksFrom.containsKey(recipient)) {
+            var whitelist = playerAllowBookmarksFrom.get(recipient);
+
+            // If target player has not added this sender to their whitelist, return early.
+            if (!whitelist.isEmpty() && !whitelist.contains(sender)) {
+                log().debug("Player " + recipient + " is not accepting bookmarks from " + sender);
                 return;
             }
+        }
 
-            if (playerAllowBookmarksFrom.containsKey(recipient)) {
-                var whitelist = playerAllowBookmarksFrom.get(recipient);
+        var players = server.getPlayerList();
+        var targetPlayer = players.getPlayer(recipient);
+        if (targetPlayer == null) return;
 
-                // If target player has not added this sender to their whitelist, return early.
-                if (!whitelist.isEmpty() && !whitelist.contains(sender)) {
-                    log().debug("Player " + recipient + " is not accepting bookmarks from " + sender);
-                    return;
-                }
-            }
-
-            var players = server.getPlayerList();
-            var targetPlayer = players.getPlayer(recipient);
-            if (targetPlayer == null) return;
-
-            log().debug("Passed checks. " + sender + " will send bookmark to target " + recipient);
-            S2CSendBookmarkToPlayer.send(targetPlayer, bookmark, photo, sender);
-        });
+        log().debug("Passed checks. " + sender + " will send bookmark to target " + recipient);
+        S2CSendBookmarkToPlayer.send(targetPlayer, bookmark, photo, sender);
     }
 
-    public void handlePlayerSettingsPacket(C2SPlayerSettings packet, ServerPlayNetworking.Context context) {
-        var player = context.player();
-        var server = player.server;
-
-        server.execute(() -> {
-            var uuid = player.getUUID();
-            playerCanReceiveBookmarks.put(uuid, packet.allowReceiving());
-            playerAllowBookmarksFrom.put(uuid, packet.allowFrom());
-        });
+    public void handlePlayerSettingsPacket(Player player, C2SPlayerSettings payload) {
+        var uuid = player.getUUID();
+        playerCanReceiveBookmarks.put(uuid, payload.allowReceiving());
+        playerAllowBookmarksFrom.put(uuid, payload.allowFrom());
     }
 }