+++ src/main/java/svenhjol/charmony/travel_journal/common/features/travel_journal/Registers.java
@@ -1,28 +1,41 @@
 package svenhjol.charmony.travel_journal.common.features.travel_journal;
 
-import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.sounds.SoundEvent;
 import svenhjol.charmony.core.base.Setup;
-import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.*;
+import svenhjol.charmony.core.common.CommonRegistry;
+import svenhjol.charmony.core.enums.Side;
+import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.C2SPlayerSettings;
+import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.C2SSendBookmarkToPlayer;
+import svenhjol.charmony.travel_journal.common.features.travel_journal.Networking.S2CSendBookmarkToPlayer;
+
+import java.util.function.Supplier;
 
 public final class Registers extends Setup<TravelJournal> {
+    public Supplier<SoundEvent> interactSound;
+    public Supplier<SoundEvent> photoSound;
+
     public Registers(TravelJournal feature) {
         super(feature);
+        var registry = CommonRegistry.forFeature(feature);
+
+        interactSound = registry.sound("interact");
+        photoSound = registry.sound("photo");
     }
 
     @Override
     public Runnable boot() {
         return () -> {
-            // Server-to-client packets
-            PayloadTypeRegistry.playS2C().register(S2CSendBookmarkToPlayer.TYPE, S2CSendBookmarkToPlayer.CODEC);
+            var registry = CommonRegistry.forFeature(feature());
+            // Server-to-client packets.
+            registry.packetSender(Side.Common, S2CSendBookmarkToPlayer.TYPE, S2CSendBookmarkToPlayer.CODEC);
 
-            // Client-to-server packets
-            PayloadTypeRegistry.playC2S().register(C2SSendBookmarkToPlayer.TYPE, C2SSendBookmarkToPlayer.CODEC);
-            PayloadTypeRegistry.playC2S().register(C2SPlayerSettings.TYPE, C2SPlayerSettings.CODEC);
+            // Client-to-server packets.
+            registry.packetSender(Side.Client, C2SSendBookmarkToPlayer.TYPE, C2SSendBookmarkToPlayer.CODEC);
+            registry.packetSender(Side.Client, C2SPlayerSettings.TYPE, C2SPlayerSettings.CODEC);
 
-            // Handle packets being sent from the client
-            ServerPlayNetworking.registerGlobalReceiver(C2SSendBookmarkToPlayer.TYPE, feature().handlers::handleSendBookmarkToPlayerPacket);
-            ServerPlayNetworking.registerGlobalReceiver(C2SPlayerSettings.TYPE, feature().handlers::handlePlayerSettingsPacket);
+            // Client packet receivers.
+            registry.packetReceiver(C2SSendBookmarkToPlayer.TYPE, () -> feature().handlers::handleSendBookmarkToPlayerPacket);
+            registry.packetReceiver(C2SPlayerSettings.TYPE, () -> feature().handlers::handlePlayerSettingsPacket);
         };
     }
 }